Added Christoph Data
I added Christoph's data selection, which works worse for Single subject but better for multi subject (which is what we are interested into). To avoid problems of git not allowing uploading big files, I divided the feature file into 3 blocks of 16MB each. I used 7zip to do it, but I guess any program should be able to merge them. Just remember to do it BEFORE trying to load it.
I have also added the function (which now is super simple) to load that data.