New evaluation methods and grouped testing
I included two extra measurements. meta cognitive accuracy MCA, and calibration delta (in addition to the MAE and MSE). Now all or some of this can be selected to be saved in the variable evaluation that goes into settings. Unfortunately, the new metrics only works with dataset one as I don't have the correctness of the trials for the dataset 2 (it will be super easy to update).

Also, I made it possible to group all testing data into a single task (to have temporal coherency). This can be activated or deactivated in teh settings with "merge_test" flag.