Validation functions and validation merge
I have made it possible to use as fitness function any of the fourst that we have implemented. This is set in
settings['val_method']. The main.py that I am currently updloading it tries both.
I have also made that if merge_test is on, the validation data comes from the same subjects and then merges it.