Added task selection
It can be selected in settings as 'select_tasks' True or False. It should be transparent for the rest of the algorithm. This does the selection for both single and meta. When merge_test = False, the selection is performed only once as average of the improvement for each test task. It could be done individually for each test task, but then the load would go into the ltl / single function instead of the splitting function. This solution is more elegant I think.